Renamed client module of bitwarden_passwordless sub-package
Renamed module `client` --> `backend`. The naming makes more sense
since the `BitwardenPasswordlessClient` communicates with the backend
API of Bitwarden Passwordless.dev and the other module of the sub-package
is called `frontend`.
